+++ .travis.yml
@@ -14,7 +14,7 @@ before_script:
   - composer self-update
   - composer install --no-interaction --prefer-source --dev
 
-script: ./vendor/bin/tester -c ./tests/php.ini-unix ./tests/GoogleAuthenticatorTests
+script: ./vendor/bin/tester -c ./tests/php.ini-unix ./tests/TotpAuthenticatorTests
 
 after_failure:
   - 'for i in $(find ./tests -name \*.actual); do echo "--- $i"; cat $i; echo; echo; done'
