+++ src/Oops/TotpAuthenticator/Utils/ITimeProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Oops\GoogleAuthenticator\Utils;
+namespace Oops\TotpAuthenticator\Utils;
 
 
 interface ITimeProvider
