+++ tests/TotpAuthenticatorTests/verifyCodeTest.phpt
@@ -1,7 +1,7 @@
 <?php
 
-use Oops\GoogleAuthenticator\Security\GoogleAuthenticator;
-use Oops\GoogleAuthenticator\Utils\ITimeProvider;
+use Oops\TotpAuthenticator\Security\TotpAuthenticator;
+use Oops\TotpAuthenticator\Utils\ITimeProvider;
 use Tester\Assert;
 
 
@@ -27,21 +27,21 @@ class TestTimeProvider implements ITimeProvider
 $seed = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ234567';
 $timeProvider = (new TestTimeProvider)->setMicroTime(1415778000);
 
-$googleAuthenticator = (new GoogleAuthenticator($timeProvider))->setTimeWindow(0);
+$googleAuthenticator = (new TotpAuthenticator($timeProvider))->setTimeWindow(0);
 Assert::false($googleAuthenticator->verifyCode('209170', $seed)); // offset = -2
 Assert::false($googleAuthenticator->verifyCode('101895', $seed)); // offset = -1
 Assert::true($googleAuthenticator->verifyCode('292224', $seed)); // offset = 0
 Assert::false($googleAuthenticator->verifyCode('800413', $seed)); // offset = +1
 Assert::false($googleAuthenticator->verifyCode('223013', $seed)); // offset = +2
 
-$googleAuthenticator = (new GoogleAuthenticator($timeProvider))->setTimeWindow(1);
+$googleAuthenticator = (new TotpAuthenticator($timeProvider))->setTimeWindow(1);
 Assert::false($googleAuthenticator->verifyCode('209170', $seed)); // offset = -2
 Assert::true($googleAuthenticator->verifyCode('101895', $seed)); // offset = -1
 Assert::true($googleAuthenticator->verifyCode('292224', $seed)); // offset = 0
 Assert::true($googleAuthenticator->verifyCode('800413', $seed)); // offset = +1
 Assert::false($googleAuthenticator->verifyCode('223013', $seed)); // offset = +2
 
-$googleAuthenticator = (new GoogleAuthenticator($timeProvider))->setTimeWindow(2);
+$googleAuthenticator = (new TotpAuthenticator($timeProvider))->setTimeWindow(2);
 Assert::true($googleAuthenticator->verifyCode('209170', $seed)); // offset = -2
 Assert::true($googleAuthenticator->verifyCode('101895', $seed)); // offset = -1
 Assert::true($googleAuthenticator->verifyCode('292224', $seed)); // offset = 0
